requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Pillow==9.2.0
torch==1.12.1
torchvision==0.13.1